Cleanup and format gl_util + window_dbg
Change size types in GL VBO type.

+++ src/gl/gl_vbo.h
@@ -5,12 +5,12 @@ struct DrawVBO {
     uint32_t vaoId      = 0;
     uint32_t vboVertId  = 0;
     uint32_t vboIdxId   = 0;
-    uint32_t nIndices   = 0;
-    int32_t vboVertSize = 0;
-    int32_t vboIdxSize  = 0;
+    int64_t nIndices   = 0;
+    int64_t vboVertSize = 0;
+    int64_t vboIdxSize  = 0;
     ~DrawVBO();
     void destroy();
     void genBuffers();
 
-    void uploadBuffer(uint32_t bufferType, void* ptr, size_t len);
+    void uploadBuffer(uint32_t bufferType, void* ptr, int64_t len);
 };
